agent/requirements.txt
websockets==14.1
openai==1.58.1
pyyaml==6.0.2
psutil==6.1.1
pyautogui==0.9.54
comtypes==1.4.8
pycaw==20240210
pywin32==308
httpx==0.27.0
